Initial skeleton for ServiceWorker
For service worker's initiation and registering code
where we prepared our code to intercept requests, it
seems like we have to run it outside of react and
redux.

For now, I put it in src/static because it has to be
loaded first. This is only the first experimental
skeleton, we might need to move it onto react later.

Note: if we take this approach, we might need to add
tests for them, a potentiall separate system from
what we have now.